Rework controller interface, add separate pitch controller

+++ src/restruct_poc/model/model.hpp
@@ -48,7 +48,7 @@ struct Model {
         return this->constraints.back();
     }
     const Constraint AddRotationControl(
-        const Node node1, const Node node2, const Array_3 axis, double* control
+        const Node node1, const Node node2, const Array_3 axis, float* control
     ) {
         this->constraints.push_back(Constraint(
             this->constraints.size(), node1, node2, ConstraintType::RotationControl, axis, control
